Optimizations, fixes, translation updates
git-svn-id: svn://svn.dolezel.info/fatrat/trunk/fatrat@45 78086a28-8538-0410-bde1-fa2ac3882955

+++ fatrat.cpp
@@ -247,6 +247,8 @@ QString formatTime(qulonglong inval)
 bool recursiveRemove(QString what)
 {
 	qDebug() << "recursiveRemove" << what;
+	if(!QFile::exists(what))
+		return true; // silently ignore
 	if(!QFile::remove(what))
 	{
 		QDir dir(what);
